+++ timesketch/frontend/src/components/Sketch/AggregationListDropdown.vue
