+++ timesketch/frontend/src/components/Sketch/EventListCompact.vue
@@ -20,7 +20,7 @@ limitations under the License.
 </template>
 
 <script>
-import ApiClient from '../../../utils/RestApiClient'
+import ApiClient from '../../utils/RestApiClient'
 import TsSketchExploreEventListItem from './EventListItem'
 
 export default {
