+++ timesketch/frontend/src/components/Sketch/SketchMetrics.vue
