+++ timesketch/frontend/src/components/Sketch/SketchSummary.vue
@@ -21,7 +21,7 @@ limitations under the License.
 </template>
 
 <script>
-import ApiClient from '../../../utils/RestApiClient'
+import ApiClient from '../../utils/RestApiClient'
 
 export default {
   name: 'ts-sketch-summary',
