+++ timesketch/frontend/src/components/Sketch/TimelinePickerItem.vue
