+++ timesketch/frontend/src/components/Sketch/UploadForm.vue
@@ -51,7 +51,7 @@ limitations under the License.
 </template>
 
 <script>
-import ApiClient from '../../../utils/RestApiClient'
+import ApiClient from '../../utils/RestApiClient'
 
 export default {
   name: 'ts-sketch-upload-timeline-form',
