+++ timesketch/frontend/src/views/SketchManage.vue
