+++ timesketch/frontend/src/views/SketchStory.vue
